Create try_diff_r_and_v
I think I am getting r and v of the air confused with r and v of the object. In here I'm trying to resolve that and do a massive overhaul on the code. This is supposed to be the same as the branch I made but I wasn't sure which was better and got confused.